# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/invitations.list": {
      "get": {
        "operationId": "InvitationsController_list",
        "parameters": [
          {
            "name": "workspaceId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of pending invitations"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "List workspace invitations",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/invitations.create": {
      "post": {
        "operationId": "InvitationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Invitation created and email sent"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Send invitation email",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/invitations.resend": {
      "post": {
        "operationId": "InvitationsController_resend",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation email resent"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Resend invitation email",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/invitations.revoke": {
      "post": {
        "operationId": "InvitationsController_revoke",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation revoked"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Revoke pending invitation",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/invitations.get": {
      "get": {
        "operationId": "InvitationsController_get",
        "parameters": [
          {
            "name": "token",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Invitation details"
          }
        },
        "security": [
          {}
        ],
        "summary": "Get invitation details by token",
        "tags": [
          "invitations"
        ]
      }
    },
    "/api/invitations.accept": {
      "post": {
        "operationId": "InvitationsController_accept",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AcceptInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Invitation accepted"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {}
        ],
        "summary": "Accept invitation",
        "tags": [
          "invitations"
        ]
      }
    },
...
    "/api/filters.list": {
      "get": {
        "operationId": "FiltersController_list",
        "parameters": [
          {
            "name": "workspace_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "required": false,
            "in": "query",
            "description": "Filter by tags",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of filters"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "List filters for workspace",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.get": {
      "get": {
        "operationId": "FiltersController_get",
        "parameters": [
          {
            "name": "workspace_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filter definition"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get filter by ID",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.create": {
      "post": {
        "operationId": "FiltersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFilterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created filter"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Create filter",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.update": {
      "post": {
        "operationId": "FiltersController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateFilterDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated filter"
          },
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Update filter",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.delete": {
      "post": {
        "operationId": "FiltersController_delete",
        "parameters": [
          {
            "name": "workspace_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Filter deleted"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Delete filter",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.reorder": {
      "post": {
        "operationId": "FiltersController_reorder",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReorderFiltersDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Filters reordered"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Reorder filters",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.listTags": {
      "get": {
        "operationId": "FiltersController_listTags",
        "parameters": [
          {
            "name": "workspace_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of unique tags"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "List unique tags across all filters",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.backfillStart": {
      "post": {
        "operationId": "FiltersController_backfillStart",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartBackfillDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Task created"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Start background backfill for all filters",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.backfillStatus": {
      "get": {
        "operationId": "FiltersController_backfillStatus",
        "parameters": [
          {
            "name": "task_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task progress"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get backfill task status",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.backfillCancel": {
      "post": {
        "operationId": "FiltersController_backfillCancel",
        "parameters": [
          {
            "name": "task_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task cancelled"
          },
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Cancel running backfill task",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.backfillList": {
      "get": {
        "operationId": "FiltersController_backfillList",
        "parameters": [
          {
            "name": "workspace_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of tasks"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "List backfill tasks for workspace",
        "tags": [
          "filters"
        ]
      }
    },
    "/api/filters.backfillSummary": {
      "get": {
        "operationId": "FiltersController_backfillSummary",
        "parameters": [
          {
            "name": "workspace_id",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Backfill summary"
          }
        },
        "security": [
          {
            "jwt-auth": []
          }
        ],
        "summary": "Get backfill status summary for workspace",
        "tags": [
          "filters"
        ]
      }
    },
...
  },
  "info": {
    "title": "Staminads API",
    "description": "Web analytics platform for tracking TimeScore metrics",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      },
      "demo-secret": {
        "type": "apiKey",
        "in": "query",
        "name": "secret",
        "description": "Demo secret for demo endpoints"
      }
    },
    "schemas": {
...
      "CreateInvitationDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "editor",
              "viewer"
            ]
          }
        },
        "required": [
          "workspace_id",
          "email",
          "role"
        ]
      },
      "AcceptInvitationDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 100
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "maxLength": 72
          }
        },
        "required": [
          "token"
        ]
      },
...
      "FilterConditionDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "enum": [
              "utm_source",
              "utm_medium",
              "utm_campaign",
              "utm_term",
              "utm_content",
              "utm_id",
              "utm_id_from",
              "referrer",
              "referrer_domain",
              "referrer_path",
              "is_direct",
              "landing_page",
              "landing_domain",
              "landing_path",
              "path",
              "device",
              "browser",
              "browser_type",
              "os",
              "user_agent",
              "connection_type",
              "language",
              "timezone"
            ]
          },
          "operator": {
            "type": "string",
            "enum": [
              "equals",
              "not_equals",
              "contains",
              "not_contains",
              "is_empty",
              "is_not_empty",
              "regex"
            ]
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "field",
          "operator"
        ]
      },
      "FilterOperationDto": {
        "type": "object",
        "properties": {
          "dimension": {
            "type": "string",
            "enum": [
              "channel",
              "channel_group",
              "stm_1",
              "stm_2",
              "stm_3",
              "stm_4",
              "stm_5",
              "stm_6",
              "stm_7",
              "stm_8",
              "stm_9",
              "stm_10",
              "utm_source",
              "utm_medium",
              "utm_campaign",
              "utm_term",
              "utm_content",
              "referrer_domain",
              "is_direct"
            ]
          },
          "action": {
            "type": "string",
            "enum": [
              "set_value",
              "unset_value",
              "set_default_value"
            ]
          },
          "value": {
            "type": "string"
          }
        },
        "required": [
          "dimension",
          "action"
        ]
      },
      "CreateFilterDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "priority": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conditions": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilterConditionDto"
            }
          },
          "operations": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilterOperationDto"
            }
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "workspace_id",
          "name",
          "conditions",
          "operations"
        ]
      },
      "UpdateFilterDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "priority": {
            "type": "number",
            "minimum": 0,
            "maximum": 1000
          },
          "order": {
            "type": "number"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "conditions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilterConditionDto"
            }
          },
          "operations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FilterOperationDto"
            }
          },
          "enabled": {
            "type": "boolean"
          }
        },
        "required": [
          "workspace_id",
          "id"
        ]
      },
      "ReorderFiltersDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "filter_ids": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "workspace_id",
          "filter_ids"
        ]
      },
      "StartBackfillDto": {
        "type": "object",
        "properties": {
          "workspace_id": {
            "type": "string"
          },
          "lookback_days": {
            "type": "number",
            "minimum": 1,
            "maximum": 365
          },
          "chunk_size_days": {
            "type": "number",
            "minimum": 1,
            "maximum": 30
          },
          "batch_size": {
            "type": "number",
            "minimum": 100,
            "maximum": 50000
          }
        },
        "required": [
          "workspace_id",
          "lookback_days"
        ]
      },
...
    }
  }
}